Wait jaeger-operator only if not installed from operatorhub

+++ install_ocp.sh
@@ -327,11 +327,15 @@ set -e
 jaeger_enabled=$(oc get syndesis app -o jsonpath='{.spec.addons.jaeger.enabled}')
 check_error $jaeger_enabled
 if [ "$jaeger_enabled" == "true" ]; then
-    wait_for sa jaeger-operator
-    result=$(oc secrets link jaeger-operator syndesis-pull-secret --for=pull)
-    check_error $result
-    # workaround as the previous "oc secrets link" doesn't trigger a pod restart
-    oc delete `oc get -o name pod -l name=jaeger-operator`
+    # on OCP 4 the jaeger-operator is installed from operatorhub on openshift-operator namespace
+    # in that case there is no need to set secrets for it
+    if [ $(is_ocp3) ]; then
+        wait_for sa jaeger-operator
+        result=$(oc secrets link jaeger-operator syndesis-pull-secret --for=pull)
+        check_error $result
+        # workaround as the previous "oc secrets link" doesn't trigger a pod restart
+        oc delete `oc get -o name pod -l name=jaeger-operator`
+    fi
 
     wait_for sa syndesis-jaeger-ui-proxy
     result=$(oc secrets link syndesis-jaeger-ui-proxy syndesis-pull-secret --for=pull)
